// src/models/message/message-content.ts
export enum MessageContent {
    Animation = 'messageAnimation',
    Audio = 'messageAudio',
    BasicGroupChatCreate = 'messageBasicGroupChatCreate',
    Call = 'messageCall',
    ChatAddMembers = 'messageChatAddMembers',
    ChatChangePhoto = 'messageChatChangePhoto',
    ChatChangeTitle = 'messageChatChangeTitle',
    ChatDeleteMember = 'messageChatDeleteMember',
    ChatDeletePhoto = 'messageChatDeletePhoto',
    ChatJoinByLink = 'messageChatJoinByLink',
    ChatSetTtl = 'messageChatSetTtl',
    ChatUpgradeFrom = 'messageChatUpgradeFrom',
    ChatUpgradeTo = 'messageChatUpgradeTo',
    Contact = 'messageContact',
    ContactRegistered = 'messageContactRegistered',
    CustomServiceAction = 'messageCustomServiceAction',
    Document = 'messageDocument',
    ExpiredPhoto = 'messageExpiredPhoto',
    ExpiredVideo = 'messageExpiredVideo',
    Game = 'messageGame',
    GameScore = 'messageGameScore',
    Invoice = 'messageInvoice',
    Location = 'messageLocation',
    PassportDataReceived = 'messagePassportDataReceived',
    PassportDataSent = 'messagePassportDataSent',
    PaymentSuccessful = 'messagePaymentSuccessful',
    PaymentSuccessfulBot = 'messagePaymentSuccessfulBot',
    Photo = 'messagePhoto',
    PinMessage = 'messagePinMessage',
    Poll = 'messagePoll',
    ScreenshotTaken = 'messageScreenshotTaken',
    Sticker = 'messageSticker',
    SupergroupChatCreate = 'messageSupergroupChatCreate',
    Text = 'messageText',
    Unsupported = 'messageUnsupported',
    Venue = 'messageVenue',
    Video = 'messageVideo',
    VideoNote = 'messageVideoNote',
    VoiceNote = 'messageVoiceNote',
    WebsiteConnected = 'messageWebsiteConnected'
}